backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "gtfs-realtime-bindings>=2.0.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
